# Single tokio runtime reuse, GIL management