Significant performance improvement when utilizing rust_bindings (Raspberry Pi 4B)

Before:
run_policy: mean=4090.6µs std=532.4µs min=3742.9µs max=6908.1µs p95=5438.6µs p99=5834.9µs (540 samples)

After:
run_policy: mean=2131.9µs std=436.2µs min=1852.2µs max=4026.2µs p95=3052.8µs p99=3764.7µs (407 samples)