@codekansas 2025-06-19
oh, awesome. great